[TASK] PHP 5.3 compatibility for Behat code
This removes the use of traits from the Behat helper code in the Flow package,
to allow the (regular) functional tests to run on PHP 5.3 while keeping the
Behat tests usable on higher PHP versions.
